// src/types/result-broadcast-tx.ts
// Generated by https://quicktype.io

/** BroadcastTx Result, compatible with `Sync` and `Async` */
export interface ResultBroadcastTxAsync {
  code: number;
  data: string;
  log: string;
  hash: string;
}

/** BroadcastTx Result */
export interface ResultBroadcastTx {
  hash: string;
  check_tx?: ResultTx;
  deliver_tx?: ResultTx;
  height?: number;
}

/** BroadcastTx inner result */
export interface ResultTx {
  code: number;
  data: string;
  log: string;
  gas_used: number;
  gas_wanted: number;
  info: string;
  tags: string[];
}

export function newResultBroadcastTx(
  hash: string,
  check_tx?: ResultTx,
  deliver_tx?: ResultTx,
  height?: number
): ResultBroadcastTx {
  return {
    hash: hash,
    check_tx: check_tx,
    deliver_tx: deliver_tx,
    height: height,
  };
}